Setting prettier and eslint
eslint prettier 설정 및 추가
