+++ .eslintrc.json
@@ -1,0 +1,13 @@
+{
+    "extends": ["react-app", "eslint:recommended", "plugin:react/recommended"],
+    "parserOptions": {
+        "ecmaVersion": 6,
+        "sourceType": "module",
+        "ecmaFeatures": {
+            "jsx": true
+        }
+    },
+    "parser": "babel-eslint",
+    "plugins": ["a-plugin"],
+    "processor": "a-plugin/a-processor"
+}
